+++ scripts/train/debug/grpo_fast.sh
@@ -2,7 +2,7 @@
 export VLLM_ALLOW_INSECURE_SERIALIZATION=1
 export VLLM_DISABLE_COMPILE_CACHE=1
 export VLLM_USE_V1=1
-uv run --active python open_instruct/grpo_fast.py \
+uv run python open_instruct/grpo_fast.py \
     --dataset_mixer_list ai2-adapt-dev/rlvr_gsm8k_zs 64 \
     --dataset_mixer_list_splits train \
     --dataset_mixer_eval_list ai2-adapt-dev/rlvr_gsm8k_zs 16 \
@@ -13,7 +13,7 @@ uv run --active python open_instruct/grpo_fast.py \
     --per_device_train_batch_size 1 \
     --num_unique_prompts_rollout 8 \
     --num_samples_per_prompt_rollout 4 \
-    --model_name_or_path deepseek-ai/DeepSeek-R1-Distill-Qwen-1.5B \
+    --model_name_or_path Qwen/Qwen3-0.6B \
     --stop_strings "</answer>" \
     --apply_verifiable_reward true \
     --temperature 0.7 \
@@ -25,11 +25,16 @@ uv run --active python open_instruct/grpo_fast.py \
     --num_epochs 1 \
     --num_learners_per_node 1 \
     --vllm_tensor_parallel_size 1 \
-    --vllm_num_engines 1 \
     --beta 0.01 \
     --seed 3 \
     --local_eval_every 1 \
+    --vllm_sync_backend gloo \
+    --vllm_gpu_memory_utilization 0.3 \
+    --save_traces \
+    --vllm_enforce_eager \
     --gradient_checkpointing \
+    --single_gpu_mode \
     --push_to_hub false \
     --system_prompt_override_file scripts/train/debug/cute_debug_system_prompt.txt \
     --active_sampling --async_steps 8
+    # --with_tracking